+++ src/App.js
@@ -1,4 +1,4 @@
-import Navigation from './components/Navigation';
+import Navigation from './components/Header';
 
 function App() {
   return (
